Moving AbbFabric and Memcached providers into separate projects
This way the base repository is not dependent on those assemblies and
can be used with inmemory or no caching providers out of the box.  Then
the separate dlls can be added for memcached or whatever you want to
use.

Also started messing with NuGet stuff so some smaple test packages are
in the source now.

References:

http://docs.nuget.org/docs/creating-packages/creating-and-publishing-a-package

http://stackoverflow.com/questions/10240029/how-to-install-a-nuget-package-nupkg-file-locally