Added Windows Azure Shared Caching provider
It seems like this is the same code as the AppFabric stuff so I'm a bit
confused.  Again haven't tested it yet.

Seems like the AppFabric stuff was used as the start for the azure
caching so they share things, but the code really is identical even
though I included different NuGet packages for them.

http://msdn.microsoft.com/en-us/library/windowsazure/gg185678.aspx