Created power shell script to automate nuget builds
It's convention based.  It goes through each sub-directory in /nuget/
and updates the version in the nuspec file based on the nuspec.template,
then uses that new file to create a new package and puts that file into
the proper nuget directory.  Then we will upload through the website as
normal.