Initial NHibernateRepository
Using NHibernate.Linq for linq syntax.  No testing done yet just getting
it to compile.  Will need to research the best practice for NHibernate
and using session.BeginTransaction() (am not using now but seems like it
should be used around the CRUD operations) and how that will play with
BatchMode.