Initial implementation for CouchDb
Not all tests pass, having an issue with PK stuff

The BaseQuery() is terrible, it brings back all records into a local
array then makes it Queryable() so it can do the linq stuff.  Obviously
this is not good.  Just trying to get it to work then figure out a way
to make that better.

Using RedBranch.Hammock (http://code.google.com/p/relax-net/) and it
references and older verison of JSON.NET so in the Integration tests I
had to add a runtime assembly binding directive to have the old version
reference actually use the newer one that is included because of a
different project refence.