More work last night, switched CouchDb C# wrapper
All of the CouchDb C# wrappers out there use a base class or a object
wrapper around your entity so that it can track id and revision in that
object.  So that didn't work because we can't make people have to have
their entity models inherit from a base class since we want themto be
POCOs.

So I took SharpCouch
(http://code.google.com/p/couchbrowse/source/browse/trunk/SharpCouch/SharpCouch.cs)
as my base and then tweaked to give it state to the server and database,
added JSON.net logic to serialize and deserialize to the generic entity
type of the repository.  Currently using a terrible BaseQuery() for
queries since it pulls back all object ids and then loads them 1 at a
time.  Just trying to get it working then will improve.  Maybe do a Linq
provider like Divian has.

Most of the integration tests pass at this point, except for the Delete
and Updates.  Not sure why yet.