Got all CouchDb integration tests to pass
There are some non-optimal solutions to some problems that come up
because CouchDb tracks an internal _id and _rev (revision) and the
revision is neded for updates or deletes, so I'm doing a separate call
to get that before processing the update or delete.  Not ideal, but
wanted to get everything working and passing then can refactor later to
make it better.

CouchDB is an easy local install
(http://wiki.apache.org/couchdb/Installing_on_Windows) if you want to
check it out.