CouchDb linq provider beginnings
Started the linq provider for CouchDB.  My approach is to use the
_temp_view functionality to provider a map function (which is just
javascript) and translate the predicate expression from LINQ to a normal
javascript if statement to take care of the querying part.  The
pagination and sorting will be tougher I think, but one thing at a time.

Also, made the GetAllDocuments() call much better.  Instead of getting
the ids and then loading them each one at a time with a call to the API,
I'm using a temp view to bring back all the document information in a
single call.