COuldn't go to sleep just yet, I will win this battle!
Made some tweaks to the linq provider and got all the integration tests
passing except for the ones that involve pagination.  The sorting is
working (limited to a single sort column which I think will be a CouchDb
restriction), but pagination will be harder since CouchDb does not
provide a way to pass in a skip param.  You can take a certain number
but that is starting from a specific key value (not an index) so will
need to find a creative way around that.

I'm sure that not all needed linq methods, etc. are implemented but
enough to pass our integration tests which is the first step I guess.
We will need to broaden the LINQ queries in the tests to cover more LINQ
issues since we are building our own now for certain implemenations.