Got all tests to pass and pagination working
Yeah me ;)

The pagination is not ideal.  Since CouchDb has a limit parameter (think
Take) but no skip parameter, instead they have a startkey parameter
which needs to take a key value and not an index.  In the CouchDb world
for pagination they suggest when showing the first 5 items (1 page) you
should actually request 6 and store the key of the 6th to use as the
startkey for page 2.  Maybe it's possible to do this but I didn't want
to deal with storing that somewhere and making sure it's for the same
previous query, etc.  Seems like a pain.

So for now for a page 2 request, I am selecting 10 items from CouchDb
and then just returning the last 5.  So it's not perfect and selects too
many items, but it works.

Still need to look into only bring back properties that are requested in
a Select statement instead of the entire object, that's next.