Added test for anonymous type in the selector
I also noticed that the selector is being applied after the results are
coming back from the server so we aren't getting the benefit of
returning less data over the wire.  We need to fix this in general but
it will take a few changes and should be done on the other branch.  So
I'll add that as an issue on the Wiki on github.