Got the pagination with selector working
Needed to implement a check for a SubQueryExpression within the
FromExpression and get the take and skip from in there since it is the
subquery for the overal Select() query that is added to the end.

Added some more tests for FindAll that use various linq syntaxes with
different data types to be able to test the linq provider better.  Right
now checks on Dates and using .Any() in the where clause are both
failing.