Fixed an issue with loading the caching strategy from configuraiton
It wasn't updating the QueryManager so it was using the
NoCachingStrategy even when the CachingStrategy got updated, the
QueryManager didn't.

Changed exceptions in configuraiton loading to
ConfigurationErrorsException instead of ArgumentException.

Added web.config transformation files but haven't got them working yet
when creating a new nuget package, that comes next.