More flexibility in using the RepositoryFactory
Now you can either load it up from the configuration file, or you can
pass in a configuration object that is a similar structure (list of
repositories, strategies and providers),  The configuration section and
the configuration object both use the same interface so the logic for
loading the repository is using that interface.

This is how enyim's memcached configuration works and I think it
provides osme nice options.