SharpRepositoryConfigurationAttribute
If you mark a class with this attribute you can give it the name of a
repository name from the configuration to use specifically for this
class.  This allows you to hook up all your classes via SM to use the
Ef5Repository, let's say, and then hook up the class SearchItem to use
RavenDbRepository so your IRepository<SearchItem, string> will get
populated from SM with a RavenDbRepositoruy while all the rest will use
the Ef5Repository.  Adds another way of configuring and overriding when
using IOC.