EF5 spike integration for compound key
I made a one off integration test for EF5 since it's not easy to
integrate this in to the other integration tests at this point because
the difference in the interfaces used.  Hopefully we can figure a way to
clean this up a bit.  But overall the POC is working I think.

Still need implementations for the real caching strategies, that comes
next.