Added timeout and standard caching strategies
Needed to make a StandardCompoundKeyCachingStrategy because there was
already a version with a 3rd generic for the Partition.

We may want to create a StandardCachingStrategyWithPartition instead so
it's obvious what it is and then we can have a
StandardCachingStrategy<T, TKey, TKey2>.  I'll probably do that later
but it will be a breaking change so need to think it thry fully first.