PoserShell script to clean duplicate entries w/ nuget
When NuGet does an update, if you have changed the config file entries
from what they are when initially installed (for example, changing the
connection string for a EF5 repository, or changing the timeout value
for the TimeoutCachingStrategy) then NuGet adds a new entry at the
bottom of the list.  The problem is that the name is the key and then
the project errors out because you can't have duplicate keys in a
section.  So the Install.ps1 script checks the web.config and/or the
app.config file to see if there are duplciate entries and if so removes
the last ones since they are the dups that were just entered.

Fixes #63