Detect circular reference in value injecter logic
When caching an object that is a DynamicProxy, we want to clean it up to
it's real base class.  Problem is we could have circular references like
UserConfiguration -> User -> UserConfiguration

So we need to keep track of the objects we've cleaned up and if we see
them again then ignore them the second time.  This appears to work, at
least in my brief testing, though much more testing is needed.