DynamicProxy integration test
Apparently the DynamicProxies for lazy loading don't happen in SqlServer
CE it seems.  So I created a testing regular SQL database to run some
tests against.

I added a SQL script to use to populate the databse.  Just create a
local SharpRepositoryTest database and run the script against it.

Got test passing.  Using caching and getting an already cached item out
and updating and saving it.  Not throwing an exception now because I'm
caching the POCO class and not the Dynamic Proxy.

I also removed some old EF bits and the old EfRepository.