Moved CouchDb into master and upgraded RavenDb.Client
Moved the CouchDbRepository project on to the master branch since it
passes all the integration tests.  It's not yet ready for prime time so
we won't make a NuGet package yet, but it's working well it seems.  I'm
sure the LINQ provider will need to be expanded to cover other things as
we move forward with it, but the basics seem to be covered.

Upgraded the RavenDB Client to the latest.