Fixed incorrect version of EF for .NET 4.0
All projects are .NET 4.0 and EF5 actually uses EF version 4.4 when EF5
is used for .NET 4.0 (confusing huh).  At some point the EF version got
switched and was using EF 5.0 even though the project was .NET 4.0.  So
this caused a problem when trying to use Ef5Repository in a .NET 4.0
project.

So this is switching it back to the correct version that will work with
.NET 4.0 and .NET 4.5 projects