ConfigurationBasedRepository and FetchStrategy update
Added new constructors for ConfigurationBasedRepository so you can pass
in the repositoryName and the configSection if they are needed, or still
use the parameterless constructor to do the defaults in the config file.

Made the logic to take the Expression based Include() parameter and
translate to a string for the EF Include statement.  Now you can do a
multiple level deep expression like x => x.OrderItems.Select(i =>
i.Product) and it will translate to "OrderItems.Product"  (thanks to
http://www.codetuning.net/blog/post/Entity-Framework-compile-safe-Includes.aspx)

Fixed #71

Also added some new FetchStrategy unit tests