Caching and Paging Options issue fixed
Fixes #74

The logic for caching TotalItems was checking for PagingOptions<T> so if
you were using the expression based PagingOptions<T, TSortKey> then it
wouldn't cache the TotalItems and then would return 0 the next time.

I added the IPagingOptions interface that both PagingOptions versions
inherit from and so I can just check for this interface and access the
TotalItems off of it.  And that fixes the issue.