Fix to allow for not defining dbContextType in config file if in the Ioc mapping
If your Ioc mapping (this is StructureMap example) defines a way to go
from a DbContext to a custom DbContext type, then you won't need to
define that in your config file since the Ioc will take care of it.
Added test coverage for this scenario and a small fix to allow it.