Added overrides for Raven, Mongo and Couch
Custom implementation of Sum for all 3.  Do a FindAll with selector to
just bring back the single column, convert to list and do Sum() on the
list.

Min and Max for Raven and Couch.  Use the PagingOptions to get back a
single record ordered by the appropriate column and return the first
item.

Not sure if Sum is the best idea to implement since it will be pretty
slow on a lot of records since the it will return all of them.  Min and
Max have don't have that side effect since we just return a single row,
so no big deal.