Changed names GroupItems is now just Group
Breaking change: removed IEnumerable<T> so you can no long do a foreach
on the repository itself.  this was a bad idea and confusing since it
exposed the IEnumerable<> methods like Aggregate, Join, GroupBy, etc. so
people may have accessed those directly which would do a GetAll() and
then do the logic on the client side