Integration test helpers
On our TeamCity build server, every once I've had to manually add the
appropriate Data directories back because they go away.  I figured an
easy solution would be to check if the directory exists and if not
create it when creating the data files for the tests.  Let's see if it
works.