Replaced Ef5Repository with EfRepository
Now that EF6 is out the version specific repository name isn't great
moving forward.  We originally had an EfRepository that was for EF4.1
and when they changed from ObjectContext to DbContext we needed to have
2 versions so we started Ef5Repository.  Now it seems EF is using the
same NuGet package with a new version so EfRepository will just require
EntityFramework 5.0 or greater.